+++ src/domain/usecases/users/UserUseCase.ts
@@ -1,7 +1,13 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  forwardRef,
+  Inject,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { KEY } from '../../models/commons/enums/Key';
 import { IBaseRepositoryAdapter } from '../../models/gateways/base/IBaseRepositoryAdapter';
 import { User } from '../../models/users/User';
+import { BookingUseCase } from '../bookings/BookingUseCase';
 import { IBaseUseCase } from '../interfaces/base/IBaseUseCase';
 
 @Injectable()
@@ -9,6 +15,8 @@ export class UserUseCase implements IBaseUseCase<User, string> {
   constructor(
     @Inject(KEY.USER_REPOSITORY)
     private readonly userRepository: IBaseRepositoryAdapter<User, string>,
+    @Inject(forwardRef(() => BookingUseCase))
+    private readonly bookingUseCase: BookingUseCase,
   ) {}
 
   public async create(data: User): Promise<User> {
@@ -20,12 +28,13 @@ export class UserUseCase implements IBaseUseCase<User, string> {
   }
 
   public async deleteById(id: string): Promise<void> {
-    try {
-      const user = await this.findById(id);
-      await this.userRepository.deleteById(user.id);
-    } catch (err) {
-      throw new Error(err);
+    const user = await this.findById(id);
+    const isUserInBooking = await this.isUserInBooking(user);
+
+    if (isUserInBooking) {
+      throw new UnauthorizedException('Cannot delete user');
     }
+    await this.userRepository.deleteById(user.id);
   }
 
   public async findAll(): Promise<User[]> {
@@ -42,5 +51,11 @@ export class UserUseCase implements IBaseUseCase<User, string> {
       throw new Error('User not found');
     }
     return user;
+  }
+
+  private async isUserInBooking(user: User): Promise<boolean> {
+    const bookings = await this.bookingUseCase.findAll();
+
+    return bookings.some((booking) => booking.userId === user.id);
   }
 }
